Only download rust toolchain once
