+++ pyodide_build/buildpkg.py
@@ -460,26 +460,6 @@ class RecipeBuilder:
         }
 
 
-def _ensure_rust_toolchain():
-    pyodide_root = get_pyodide_root()
-    toolchain_version = "emscripten_2025-01-13_052ba16"
-    toolchain_path = pyodide_root / ".rust-toolchain" / toolchain_version
-    if toolchain_path.exists():
-        return
-    from .xbuildenv import download_and_unpack_archive
-
-    download_and_unpack_archive(
-        f"http://pyodide-cache.s3-website-us-east-1.amazonaws.com/rustc/{toolchain_version}.tar.bz2",
-        toolchain_path.parent,
-    )
-    result = subprocess.run(
-        ["rustup", "toolchain", "link", toolchain_version, toolchain_path], check=False
-    )
-    if result.returncode != 0:
-        logger.error("ERROR: rustup toolchain install failed")
-        exit_with_stdio(result)
-
-
 class RecipeBuilderPackage(RecipeBuilder):
     """
     Recipe builder for python packages.
@@ -487,7 +467,6 @@ class RecipeBuilderPackage(RecipeBuilder):
 
     def _build_package(self, bash_runner: BashRunnerWithSharedEnvironment) -> None:
         if self.recipe.is_rust_package():
-            _ensure_rust_toolchain()
             bash_runner.run(
                 RUST_BUILD_PRELUDE,
                 script_name="rust build prelude",
